+++ lib/src/bones_api_entity_db_postgres.dart
@@ -170,11 +170,10 @@ class DBPostgreSQLAdapter extends DBSQLAdapter<PostgreSQLConnectionWrapper>
     minConnections ??= 1;
     maxConnections ??= 3;
 
-    var retCheckTablesAndGenerateTables =
-        DBSQLAdapter.parseConfigDBGenerateTablesAndCheckTables(config);
-
-    var generateTables = retCheckTablesAndGenerateTables[0];
-    var checkTables = retCheckTablesAndGenerateTables[1];
+    var (
+      generateTables: generateTables,
+      checkTables: checkTables,
+    ) = DBSQLAdapter.parseConfigDBGenerateTablesAndCheckTables(config);
 
     var populate = config?['populate'];
     Object? populateTables;